+++ defines/yes3_defines.php
@@ -3,10 +3,14 @@
 define('ONE_MINUTE', 60);
 define('ONE_HOUR', 60*60);
 define('ONE_DAY', 24*60*60);
-define('ALL_OF_THEM', "all");
+define('ALL_OF_THEM', "all"); // we'd better change this
 define('MAX_LABEL_LEN', 128);
 
 define('EMLOG_MSG_EXPORT_SPECIFICATION', 'yes3-export-specification');
+define('EMLOG_MSG_EXPORT_EVENTS', 'yes3-export-events');
+
+define('EMLOG_TYPE_EXPORT_SPECIFICATION', 'yes3-export-specification');
+define('EMLOG_TYPE_EXPORT_EVENTS', 'yes3-export-events');
 define('EMLOG_TYPE_EXPORT_LOG_ENTRY', 'yes3-export-log');
 define('EMLOG_TYPE_ERROR_REPORT', 'yes3-export-error-report');
 define('EMLOG_TYPE_CRON_LOG', 'yes3-exporter-cron-log');
@@ -16,4 +20,5 @@ define('VARNAME_GROUP_NAME', 'redcap_data_access_group');
 
 define('VARNAME_EVENT_ID',   'redcap_event_id');
 define('VARNAME_EVENT_NAME', 'redcap_event_name');
+
 define('VARNAME_INSTANCE',   'redcap_repeat_instance');